// src/_shared/components/storyblok/schema/component-types.d.ts
// This file was generated by the storyblok CLI.
// DO NOT MODIFY THIS FILE BY HAND.
import type { ISbStoryData } from "storyblok";
export interface AssetStoryblok {
  alt?: string;
  copyright?: string;
  id: number;
  filename: string;
  name: string;
  title?: string;
  focus?: string;
  [k: string]: any;
}

export interface CardStoryblok {
  value: string;
  picture: AssetStoryblok;
  component: "card";
  _uid: string;
  [k: string]: any;
}